+++ src/main/java/net/pcal/fastback/repo/RepoImpl.java
@@ -31,9 +31,7 @@ import org.eclipse.jgit.lib.Ref;
 import org.eclipse.jgit.util.FileUtils;
 
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.time.Duration;
@@ -142,7 +140,7 @@ class RepoImpl implements Repo {
 
     @Override
     public String getWorldUuid() throws IOException {
-        return UuidUtils.getWorldUuid(this.getWorkTree().toPath());
+        return WorldIdUtils.getWorldUuid(this.getWorkTree().toPath());
     }
 
     @Override
